[ORGA] Extract Student::Sco::HeaderActions component
