+++ orga/app/templates/authenticated/sco-students/list.hbs
@@ -1,14 +1,23 @@
 {{page-title (t 'pages.students-sco.title')}}
+
 <div class="list-students-page">
+  <Student::Sco::HeaderActions
+    @onImportStudents={{this.importStudents}}
+  />
+
   <Student::Sco::List
     @students={{@model}}
-    @isLoading={{this.isLoading}}
     @connectionTypesOptions={{this.connectionTypesOptions}}
     @lastNameFilter={{this.lastName}}
     @firstNameFilter={{this.firstName}}
     @connexionTypeFilter={{this.connexionType}}
     @onFilter={{this.triggerFiltering}}
-    @onImportStudents={{this.importStudents}}
     @onDissociateUser={{this.refresh}}
   />
+
+  {{#if this.isLoading}}
+    <ModalDialog @translucentOverlay={{true}} @containerClass="modal__translucent">
+      <PixLoader />
+    </ModalDialog>
+  {{/if}}
 </div>
